finishing mark and sweep

+++ js/main.js
@@ -1,28 +1,54 @@
 import { Simulation } from './Simulation.js';
 import { Object } from './tree-structure/Object.js';
 
-document.addEventListener("DOMContentLoaded", function(event) {
+let simulation = null;
+
+document.addEventListener('DOMContentLoaded', function(event) {
   main();
 });
 
-function main()
+document.getElementById('next').addEventListener('click', function(event) {
+  nextStep()
+});
+
+document.getElementById('reset').addEventListener('click', function(event) {
+  initSimulation()
+});
+
+function initSimulation()
 {
   let divMarkSweep = document.getElementById('mark-sweep');
   let divListObjects = document.getElementById('list-objects-container');
 
-  let simulation = new Simulation(divMarkSweep, divListObjects);
+  document.getElementById('next').disabled = false;
+
+  simulation = new Simulation(divMarkSweep, divListObjects);
 
   buildObjects(simulation);
 
   simulation.render();
   simulation.update();
 
-  let notFinished = true;
-  while(notFinished)
+  document.getElementById('action').innerHTML = '-';
+}
+
+function nextStep()
+{
+  if(!simulation.nextMarkTree())
   {
-    notFinished = simulation.nextMarkTree();
+    document.getElementById('action').innerHTML = 'Suppression des objets non-marqués';
+    simulation.deleteUnmarked();
+    document.getElementById('next').disabled = true;
   }
-  simulation.deleteUnmarked();
+  else
+  {
+    document.getElementById('action').innerHTML = 'Objet marqué';
+  }
+}
+
+function main()
+{
+  initSimulation();
 }
 
 function buildObjects(simulation)
